- Trabalho quase pronto para apresentação - Falta apenas salvar os depositos

+++ TrabHospital/TrabHospital/Modelo/Deposito.cs
@@ -10,14 +10,14 @@ namespace TrabHospital.Modelo
 	{
 		private int parcela;
 		private DateTime data, dtcompensa;
-		private float valor;
+		private double valor;
 		private String cheque;
 		private char tipo;
 		private Atendimento atendimento;
 
 		public Deposito() { }
 
-		public Deposito(int parcela, DateTime data, DateTime dtcompensa, float valor, string cheque, char tipo, Atendimento atendimento)
+		public Deposito(int parcela, DateTime data, DateTime dtcompensa, double valor, string cheque, char tipo, Atendimento atendimento)
 		{
 			this.parcela = parcela;
 			this.data = data;
@@ -31,7 +31,7 @@ namespace TrabHospital.Modelo
 		public int Parcela { get => parcela; set => parcela = value; }
 		public DateTime Data { get => data; set => data = value; }
 		public DateTime Dtcompensa { get => dtcompensa; set => dtcompensa = value; }
-		public float Valor { get => valor; set => valor = value; }
+		public double Valor { get => valor; set => valor = value; }
 		public string Cheque { get => cheque; set => cheque = value; }
 		public char Tipo { get => tipo; set => tipo = value; }
 		internal Atendimento Atendimento { get => atendimento; set => atendimento = value; }
